+++ changelog.md
@@ -1,1 +1,2 @@
 - Fixed gas pipes not working with mekanism multiblocks
+- Fixed upgrade consumed in creative